add typecasts to avoid implicit conversions

+++ include/libflatarray/detail/load_functor.hpp
@@ -42,10 +42,14 @@ public:
 
         for (ITERATOR i = start; i != end; ++i) {
             accessor.index() = soa_accessor<CELL, DIM_X, DIM_Y, DIM_Z, INDEX>::gen_index(
-                i->origin[0],
-                i->origin[1],
-                i->origin[2]);
-            accessor.load(source, i->length(), offset, count);
+                static_cast<long>(i->origin[0]),
+                static_cast<long>(i->origin[1]),
+                static_cast<long>(i->origin[2]));
+            accessor.load(
+                source,
+                i->length(),
+                static_cast<std::size_t>(offset),
+                count);
 
             offset += i->length();
         }
